kotak-algo/requirements.txt
requests
sqlalchemy
fastapi
uvicorn
websockets
pyotp
python-dotenv
pandas
apscheduler
pytz
rich
aiohttp
jinja2